new: check_kernel_status for Debian/Ubuntu
A nice little perl script that parses /proc/version and compares the
version of the current running kernel to those reported by 'dpkg -l'. It
changes to "WARNING" state when an installed kernel is newer than the
running one.

I've modified it so that it now works well with current Lenny and
Squeeze kernel packages (Squeeze+Xen also).

Signed-off-by: Gabriel Filion <lelutin@gmail.com>