New: check_debsums
Check the consistency of installed files for one or more packages.

I wrote this plugin in order to catch script kiddies on an already
infected server right when they opened their backdoor (which was in an
altered sshd executable). This way I found out how they logged in and
locked them out until I was ready to reinstall.

This plugin will compute file hashes and compare them to those from the
.deb packages (hurray for debsums!).

Since the operation is not very light, it shouldn't be used on too much
packages at a time.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>