silence annoying text to stderr when checking sigs
when we run the signature check to make sure there's no revocation sig
in place, gpg (2.1.x at least..) always outputs an annoying line on
stderr about how many valid signatures it found. we don't need to know
this since we're checking the output on stdout, and it shouldn't be
leaking out to the terminal when run manually.